+++ RenderEngine/material/DiffuseEmitter.cpp
@@ -11,31 +11,27 @@
 bool DiffuseEmitter::m_optixMaterialIsCreated = false;
 optix::Material DiffuseEmitter::m_optixMaterial;
 
-DiffuseEmitter::DiffuseEmitter(const Vector3& power, const Vector3& Kd)
-    : m_power(power)
-    , m_Kd(Kd)
-    , m_inverseArea(0)
-{
+DiffuseEmitter::DiffuseEmitter(const Vector3 &power, const Vector3 &Kd)
+        : m_power(power), m_Kd(Kd), m_inverseArea(0) {
     m_power.x *= Kd.x;
     m_power.y *= Kd.y;
     m_power.z *= Kd.z;
 }
 
-optix::Material DiffuseEmitter::getOptixMaterial(optix::Context& context)
-{
-    if (!m_optixMaterialIsCreated)
-    {
+optix::Material DiffuseEmitter::getOptixMaterial(optix::Context &context) {
+    if (!m_optixMaterialIsCreated) {
         optix::Program radianceProgram
-            = context->createProgramFromPTXFile(getPtxFile("material/DiffuseEmitter.ptx"), "closestHitRadiance");
+                = context->createProgramFromPTXFile(getPtxFile("material/DiffuseEmitter.ptx"), "closestHitRadiance");
         m_optixMaterial = context->createMaterial();
-        m_optixMaterial->setClosestHitProgram(RayType::RADIANCE, radianceProgram);
-        m_optixMaterial->setClosestHitProgram(RayType::RADIANCE_IN_PARTICIPATING_MEDIUM, radianceProgram);
+        m_optixMaterial->setClosestHitProgram(static_cast<int>(RayType::RADIANCE), radianceProgram);
+        m_optixMaterial->setClosestHitProgram(static_cast<int>(RayType::RADIANCE_IN_PARTICIPATING_MEDIUM),
+                                              radianceProgram);
         m_optixMaterial->setClosestHitProgram(
-            RayType::PHOTON,
-            context->createProgramFromPTXFile(getPtxFile("material/DiffuseEmitter.ptx"), "closestHitPhoton"));
+                static_cast<int>(RayType::PHOTON),
+                context->createProgramFromPTXFile(getPtxFile("material/DiffuseEmitter.ptx"), "closestHitPhoton"));
         m_optixMaterial->setAnyHitProgram(
-            RayType::SHADOW,
-            context->createProgramFromPTXFile(getPtxFile("material/DiffuseEmitter.ptx"), "gatherAnyHitOnEmitter"));
+                static_cast<int>(RayType::SHADOW),
+                context->createProgramFromPTXFile(getPtxFile("material/DiffuseEmitter.ptx"), "gatherAnyHitOnEmitter"));
         this->registerMaterialWithShadowProgram(context, m_optixMaterial);
         m_optixMaterialIsCreated = true;
     }
@@ -46,8 +42,7 @@ optix::Material DiffuseEmitter::getOptixMaterial(optix::Context& context)
 // Register any material-dependent values to be available in the optix program.
 */
 
-void DiffuseEmitter::registerGeometryInstanceValues(optix::GeometryInstance& instance)
-{
+void DiffuseEmitter::registerGeometryInstanceValues(optix::GeometryInstance &instance) {
     instance["Kd"]->setFloat(m_Kd);
     instance["Ks"]->setFloat(0.0f, 0.0f, 0.0f);
     optix::float3 powerPerArea = m_power;
@@ -60,12 +55,10 @@ void DiffuseEmitter::registerGeometryInstanceValues(optix::GeometryInstance& ins
     instance["power"]->setFloat(m_power);
 }
 
-Vector3 DiffuseEmitter::getPower() const
-{
+Vector3 DiffuseEmitter::getPower() const {
     return m_power;
 }
 
-void DiffuseEmitter::setInverseArea(float inverseArea)
-{
+void DiffuseEmitter::setInverseArea(float inverseArea) {
     m_inverseArea = inverseArea;
 }
