+++ RenderEngine/material/ParticipatingMedium.cpp
@@ -29,10 +29,10 @@ optix::Material ParticipatingMedium::getOptixMaterial(optix::Context& context)
         // context->createProgramFromPTXFile(getPtxFile("material/ParticipatingMedium.ptx"), "radianceTransmission");
 
         m_optixMaterial = context->createMaterial();
-        m_optixMaterial->setClosestHitProgram(RayType::RADIANCE, radianceProgram);
-        m_optixMaterial->setClosestHitProgram(RayType::RADIANCE_IN_PARTICIPATING_MEDIUM, radianceProgram);
-        m_optixMaterial->setClosestHitProgram(RayType::PHOTON, photonProgram);
-        m_optixMaterial->setClosestHitProgram(RayType::PHOTON_IN_PARTICIPATING_MEDIUM, photonProgram);
+        m_optixMaterial->setClosestHitProgram(static_cast<int>(RayType::RADIANCE), radianceProgram);
+        m_optixMaterial->setClosestHitProgram(static_cast<int>(RayType::RADIANCE_IN_PARTICIPATING_MEDIUM), radianceProgram);
+        m_optixMaterial->setClosestHitProgram(static_cast<int>(RayType::PHOTON), photonProgram);
+        m_optixMaterial->setClosestHitProgram(static_cast<int>(RayType::PHOTON_IN_PARTICIPATING_MEDIUM), photonProgram);
 
         this->registerMaterialWithShadowProgram(context, m_optixMaterial);
